Seperate task sorting and date formatting to tasksUtils.ts
