+++ frontend/src/App.tsx
@@ -3,26 +3,20 @@ import { ChakraProvider, Box } from "@chakra-ui/react";
 import Home from "./components/Home";
 import { Task } from "./types/Task";
 import { fetchTasks } from "./api/tasksApi"; 
+import { sortTasks } from "./utils/taskUtils";
 
 const App: React.FC = () => {
   const [tasks, setTasks] = useState<Task[]>([]);
 
-  const loadTasks = async () => {
-    try {
-      const fetchedTasks: Task[] = await fetchTasks();
-      fetchedTasks.sort((a, b) => {
-        const priorityOrder = { High: 1, Medium: 2, Low: 3 };
-        return (
-          priorityOrder[a.priority] - priorityOrder[b.priority] ||
-          new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime()
-        );
-      });
-
-      setTasks(fetchedTasks);
-    } catch (error) {
-      console.error("Failed to load tasks:", error);
-    }
-  };
+const loadTasks = async () => {
+  try {
+    let fetchedTasks: Task[] = await fetchTasks();
+    fetchedTasks = sortTasks(fetchedTasks); 
+    setTasks(fetchedTasks);
+  } catch (error) {
+    console.error("Failed to load tasks:", error);
+  }
+};
 
   useEffect(() => {
     loadTasks();
